tiedostoihin jako
